improve diagnostics
