+++ src/result.rs
@@ -1,0 +1,44 @@
+//! Result types to enabling handling of ODBC Errors
+use super::DiagnosticRecord;
+use std::fmt::{Display, Formatter};
+use std;
+
+pub type Result<T> = std::result::Result<T, DiagnosticRecord>;
+
+#[must_use]
+pub enum Return<T> {
+    Success(T),
+    SuccessWithInfo(T),
+    Error,
+}
+
+/// Environment allocation error
+///
+/// Allocating an environment is the one operation in ODBC which does not yiel a diagnostic record
+/// in case of an error. There is simply no Environment to ask for a diagnostic record
+#[derive(Debug)]
+pub enum EnvAllocError {
+    Info(DiagnosticRecord),
+    Error,
+}
+
+impl Display for EnvAllocError {
+    fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
+        use std::error::Error;
+        write!(f, "{}", self.description())
+    }
+}
+
+impl std::error::Error for EnvAllocError {
+    fn description(&self) -> &str {
+        match self {
+            &EnvAllocError::Info(ref dr) => dr.description(),
+            &EnvAllocError::Error => "Failure to allocate ODBC environment",
+        }
+    }
+
+    fn cause(&self) -> Option<&std::error::Error> {
+        None
+    }
+}
+
